Made jsonobject class file outside of gsclient
Not sure if this helps or not, but jsonobject is no longer a nested
class within gsclient.
