--- GUI/UAV-Forge-New/qtmapper.cpp
@@ -1,5 +1,0 @@
-#include "qtmapper.h"
-
-QtMapper::QtMapper()
-{
-}
