--- GUI/UAV-Forge-New/qtmapper.h
@@ -1,10 +1,0 @@
-#ifndef QTMAPPER_H
-#define QTMAPPER_H
-
-class QtMapper
-{
-public:
-    QtMapper();
-};
-
-#endif // QTMAPPER_H
